requirements.txt
opencv-python==4.6.0.66
numpy==1.21.6
click==8.0.1
PyYAML==6.0